refactor: add useModalAddPhoto hook
- Update generalInfo.module.scss
- Add js doc comments
